@@ -1,8 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <Version>4.1.11</Version>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <Version>4.1.10</Version>
     <Description>The storefront implementation of the Virto Commerce platform.</Description>
     <Copyright>Copyright © 2011-2019 Virto Commerce - an enterprise Microsoft ecommerce platform ©. All rights reserved</Copyright>
     <Authors>Virto Commerce</Authors>
@@ -12,17 +12,19 @@
     <PackageProjectUrl>https://github.com/VirtoCommerce/vc-storefront-core</PackageProjectUrl>
     <PackageIconUrl>https://virtocommerce.com/themes/assets/logo.jpg</PackageIconUrl>
     <RepositoryUrl>https://github.com/VirtoCommerce/vc-storefront-core</RepositoryUrl>
-    <AssemblyVersion>4.1.11.0</AssemblyVersion>
-    <FileVersion>4.1.11.0</FileVersion>
+    <AssemblyVersion>4.1.10.0</AssemblyVersion>
+    <FileVersion>4.1.10.0</FileVersion>
+    <IsPackable>False</IsPackable>
+    <OutputType>Library</OutputType>
   </PropertyGroup>
 
 
   <ItemGroup>
-    <PackageReference Include="LibSassHost" Version="1.2.10" />
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="LibSassHost" Version="1.3.0" />
     <PackageReference Include="LibSassHost.Native.linux-x64" Version="1.2.10" />
     <PackageReference Include="LibSassHost.Native.win-x64" Version="1.2.10" />
     <PackageReference Include="LibSassHost.Native.win-x86" Version="1.2.10" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
     <PackageReference Include="Nito.AsyncEx" Version="5.0.0" />
     <PackageReference Include="PagedList.Core" Version="1.17.4" />
     <PackageReference Include="Scriban" Version="2.1.2" />
