@@ -7,16 +7,16 @@ namespace VirtoCommerce.Storefront.Model
     {
         public SettingEntry()
         {
-            AllowedValues = new List<string>();
-            ArrayValues = new List<string>();
+            AllowedValues = new List<object>();
+            ArrayValues = new List<object>();
         }
         public string Name { get; set; }
-        public string Value { get; set; }
+        public object Value { get; set; }
         public string ValueType { get; set; }
-        public IList<string> AllowedValues { get; set; }
-        public string DefaultValue { get; set; }
+        public IList<object> AllowedValues { get; set; }
+        public object DefaultValue { get; set; }
         public bool IsArray { get; set; }
-        public IList<string> ArrayValues { get; set; }
+        public IList<object> ArrayValues { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
 
