@@ -5,8 +5,8 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
-using VirtoCommerce.Storefront.AutoRestClients.PlatformModuleApi;
-using VirtoCommerce.Storefront.AutoRestClients.PlatformModuleApi.Models;
+using VirtoCommerce.Storefront.AutoRestClients.NotificationsModuleApi;
+using VirtoCommerce.Storefront.AutoRestClients.NotificationsModuleApi.Models;
 using VirtoCommerce.Storefront.Domain;
 using VirtoCommerce.Storefront.Domain.Common;
 using VirtoCommerce.Storefront.Domain.Security;
@@ -876,18 +876,21 @@ namespace VirtoCommerce.Storefront.Controllers
         }
 
 
-        private async Task<SendNotificationResult> SendNotificationAsync(NotificationBase notification)
+        private async Task<NotificationSendResult> SendNotificationAsync(NotificationBase notification)
         {
-            var result = new SendNotificationResult();
+            NotificationSendResult result;
 
             try
             {
-                result = await _platformNotificationApi.SendNotificationAsync(notification.ToNotificationDto());
+                result = await _platformNotificationApi.SendNotificationByRequestAsync(notification.ToNotificationDto());
             }
             catch
             {
-                result.IsSuccess = false;
-                result.ErrorMessage = "Error occurred while sending notification";
+                result = new NotificationSendResult
+                {
+                    IsSuccess = false,
+                    ErrorMessage = "Error occurred while sending notification"
+                };
             }
 
             return result;
