@@ -1,4 +1,5 @@
 using System.Linq;
+using VirtoCommerce.Storefront.AutoRestClients.NotificationsModuleApi.Models;
 using VirtoCommerce.Storefront.AutoRestClients.PlatformModuleApi.Models;
 using VirtoCommerce.Storefront.Model.Common.Notifications;
 
@@ -6,9 +7,9 @@ namespace VirtoCommerce.Storefront.Domain.Common
 {
     public static class NotificationConverter
     {
-        public static TestNotificationRequest ToNotificationDto(this NotificationBase notification)
+        public static NotificationRequest ToNotificationDto(this NotificationBase notification)
         {
-            var result = new TestNotificationRequest
+            var result = new NotificationRequest
             {
                 Language = notification.Language.CultureName,
                 ObjectId = notification.StoreId,
