@@ -91,7 +91,7 @@ namespace VirtoCommerce.Storefront.Domain
             var existContact = await GetContactByIdAsync(contactId);
             if (existContact != null)
             {
-                await _customerApi.UpdateAddessesAsync(contactId, addresses.Select(x => x.ToCustomerAddressDto()).ToList());
+                await _customerApi.UpdateAddessesAsync(addresses.Select(x => x.ToCustomerAddressDto()).ToList(), contactId);
 
                 //Invalidate cache
                 CustomerCacheRegion.ExpireMember(existContact.Id);
@@ -211,7 +211,7 @@ namespace VirtoCommerce.Storefront.Domain
                 ObjectType = "Member"
             };
 
-            var searchResult = await _customerApi.SearchAsync(criteriaDto);
+            var searchResult = await _customerApi.SearchMemberAsync(criteriaDto);
             var contacts = _customerApi.GetContactsByIds(searchResult.Results.Select(x => x.Id).ToList()).Select(x => x.ToContact()).ToList();
 
             return new StaticPagedList<Contact>(contacts, criteria.PageNumber, criteria.PageSize, searchResult.TotalCount.Value);
