@@ -8,6 +8,7 @@ using VirtoCommerce.Storefront.Model.Stores;
 using coreDto = VirtoCommerce.Storefront.AutoRestClients.CoreModuleApi.Models;
 using platformDto = VirtoCommerce.Storefront.AutoRestClients.PlatformModuleApi.Models;
 using storeDto = VirtoCommerce.Storefront.AutoRestClients.StoreModuleApi.Models;
+using paymentDto = VirtoCommerce.Storefront.AutoRestClients.PaymentModuleApi.Models;
 
 namespace VirtoCommerce.Storefront.Domain
 {
@@ -22,7 +23,7 @@ namespace VirtoCommerce.Storefront.Domain
 
         public static DynamicProperty ToDynamicProperty(this storeDto.DynamicObjectProperty propertyDto)
         {
-            return propertyDto.JsonConvert<coreDto.DynamicObjectProperty>().ToDynamicProperty();
+            return propertyDto.JsonConvert<platformDto.DynamicObjectProperty>().ToDynamicProperty();
         }
 
         public static Store ToStore(this storeDto.Store storeDto)
@@ -76,7 +77,7 @@ namespace VirtoCommerce.Storefront.Domain
             if (!storeDto.Settings.IsNullOrEmpty())
             {
                 result.Settings = new MutablePagedList<SettingEntry>(storeDto.Settings.Where(x => !x.ValueType.EqualsInvariant("SecureString"))
-                                                                                      .Select(x => x.JsonConvert<platformDto.Setting>()
+                                                                                      .Select(x => x.JsonConvert<platformDto.ObjectSettingEntry>()
                                                                                       .ToSettingEntry()));
             }
 
@@ -93,7 +94,7 @@ namespace VirtoCommerce.Storefront.Domain
         }
 
 
-        public static PaymentMethod ToStorePaymentMethod(this storeDto.PaymentMethod paymentMethodDto, Currency currency)
+        public static PaymentMethod ToStorePaymentMethod(this paymentDto.PaymentMethod paymentMethodDto, Currency currency)
         {
             var retVal = new PaymentMethod(currency)
             {
@@ -111,7 +112,7 @@ namespace VirtoCommerce.Storefront.Domain
 
             if (paymentMethodDto.Settings != null)
             {
-                retVal.Settings = paymentMethodDto.Settings.Where(x => !x.ValueType.EqualsInvariant("SecureString")).Select(x => x.JsonConvert<AutoRestClients.PlatformModuleApi.Models.Setting>().ToSettingEntry()).ToList();
+                retVal.Settings = paymentMethodDto.Settings.Where(x => !x.ValueType.EqualsInvariant("SecureString")).Select(x => x.JsonConvert<AutoRestClients.PlatformModuleApi.Models.ObjectSettingEntry>().ToSettingEntry()).ToList();
             }
 
             retVal.Currency = currency;
@@ -127,7 +128,7 @@ namespace VirtoCommerce.Storefront.Domain
             return retVal;
         }
 
-        public static TaxDetail ToTaxDetail(this storeDto.TaxDetail dto, Currency currency)
+        public static TaxDetail ToTaxDetail(this paymentDto.TaxDetail dto, Currency currency)
         {
             var result = new TaxDetail(currency)
             {
