@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using VirtoCommerce.Storefront.AutoRestClients.CoreModuleApi;
+using VirtoCommerce.Storefront.AutoRestClients.TaxModuleApi;
 using VirtoCommerce.Storefront.Caching;
 using VirtoCommerce.Storefront.Extensions;
 using VirtoCommerce.Storefront.Model;
@@ -12,17 +13,17 @@ using VirtoCommerce.Storefront.Model.Caching;
 using VirtoCommerce.Storefront.Model.Common.Caching;
 using VirtoCommerce.Storefront.Model.Tax;
 using VirtoCommerce.Storefront.Model.Tax.Services;
-using coreService = VirtoCommerce.Storefront.AutoRestClients.CoreModuleApi.Models;
+using taxDto = VirtoCommerce.Storefront.AutoRestClients.TaxModuleApi.Models;
 
 namespace VirtoCommerce.Storefront.Domain
 {
     public class TaxEvaluator : ITaxEvaluator
     {
-        private readonly ICommerce _commerceApi;
+        private readonly ITaxModule _taxApi;
         private readonly IStorefrontMemoryCache _memoryCache;
-        public TaxEvaluator(ICommerce commerceApi, IStorefrontMemoryCache memoryCache)
+        public TaxEvaluator(ITaxModule taxApi, IStorefrontMemoryCache memoryCache)
         {
-            _commerceApi = commerceApi;
+            _taxApi = taxApi;
             _memoryCache = memoryCache;
         }
 
@@ -38,7 +39,7 @@ namespace VirtoCommerce.Storefront.Domain
             {
                 throw new ArgumentNullException(nameof(owners));
             }
-            IList<coreService.TaxRate> taxRates = new List<coreService.TaxRate>();
+            IList<taxDto.TaxRate> taxRates = new List<taxDto.TaxRate>();
             if (context.StoreTaxCalculationEnabled)
             {
                 //Do not execute platform API for tax evaluation if fixed tax rate is used
@@ -46,7 +47,7 @@ namespace VirtoCommerce.Storefront.Domain
                 {
                     foreach (var line in context.Lines ?? Enumerable.Empty<TaxLine>())
                     {
-                        var rate = new coreService.TaxRate()
+                        var rate = new taxDto.TaxRate()
                         {
                             Rate = (double)(line.Amount * context.FixedTaxRate * 0.01m).Amount,
                             Currency = context.Currency.Code,
@@ -61,7 +62,7 @@ namespace VirtoCommerce.Storefront.Domain
                     taxRates = await _memoryCache.GetOrCreateExclusiveAsync(cacheKey, (cacheEntry) =>
                     {
                         cacheEntry.AddExpirationToken(TaxCacheRegion.CreateChangeToken());
-                        return _commerceApi.EvaluateTaxesAsync(context.StoreId, context.ToTaxEvaluationContextDto());
+                        return _taxApi. EvaluateTaxesAsync(context.StoreId, context.ToTaxEvaluationContextDto());
                     });
                 }
             }
@@ -70,7 +71,7 @@ namespace VirtoCommerce.Storefront.Domain
 
         #endregion
 
-        private static void ApplyTaxRates(IList<coreService.TaxRate> taxRates, IEnumerable<ITaxable> owners)
+        private static void ApplyTaxRates(IList<taxDto.TaxRate> taxRates, IEnumerable<ITaxable> owners)
         {
             if (taxRates == null)
             {
