@@ -49,7 +49,7 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
         }
 
         /// <summary>
-        ///  Initializes a new instance of the <see cref="AccessTokenDelegatingHandler"/> class.
+        ///  Initializes a new instance of the <see cref="ClientCredentialsAuthHandler"/> class.
         /// </summary>
         /// <param name="options"></param>
         /// <param name="clientFactory"></param>
@@ -70,7 +70,7 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
         protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
             var accessToken = await GetAccessTokenAsync(cancellationToken);
-            if (string.IsNullOrWhiteSpace(accessToken) && (await RenewTokensAsync(cancellationToken) == false))
+            if (string.IsNullOrWhiteSpace(accessToken) && (!(await RenewTokensAsync(cancellationToken))))
             {
                 return new HttpResponseMessage(HttpStatusCode.Unauthorized) { RequestMessage = request };
             }
@@ -83,7 +83,7 @@ namespace VirtoCommerce.Storefront.Infrastructure.Autorest
                 return response;
             }
 
-            if (await RenewTokensAsync(cancellationToken) == false)
+            if (!(await RenewTokensAsync(cancellationToken)))
             {
                 return response;
             }
