@@ -1,28 +1,29 @@
 using System.Linq;
 using Microsoft.AspNetCore.Mvc.Controllers;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 namespace VirtoCommerce.Storefront.Infrastructure.Swagger
 {
     public class OptionalParametersFilter : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
-            if (operation.Parameters?.Any() == true)
+            if (operation.Parameters == null || !operation.Parameters.Any())
             {
-                var optionalParameters = context.ApiDescription.ParameterDescriptions
-                    .Where(x => x.ParameterDescriptor != null &&
-                        ((ControllerParameterDescriptor)x.ParameterDescriptor).ParameterInfo.CustomAttributes.Any(attr => attr.AttributeType == typeof(SwaggerOptionalAttribute)))
-                    .ToList();
+                return;
+            }
 
-                foreach (var apiParameter in optionalParameters)
+            var optionalParameters = context.ApiDescription.ParameterDescriptions
+                .Where(p => p.ParameterDescriptor != null &&
+                ((ControllerParameterDescriptor)p.ParameterDescriptor).ParameterInfo.CustomAttributes.Any(attr => attr.AttributeType == typeof(SwaggerOptionalAttribute))).ToList();
+
+            foreach (var apiParameter in optionalParameters)
+            {
+                var parameter = operation.Parameters.FirstOrDefault(p => p.Name == apiParameter.Name);
+                if (parameter != null)
                 {
-                    var parameter = operation.Parameters.FirstOrDefault(x => x.Name == apiParameter.Name);
-                    if (parameter != null)
-                    {
-                        parameter.Required = false;
-                    }
+                    parameter.Required = false;
                 }
             }
         }
